fix: make popover usable within components having focus
the 'onFocusOutside' handler of react-foco was removed. Using this trigger causes react-foco to handle the events in the capturing event phase which therefore then can't be handled by the component as the handler closes the Popover. As the Popover uses the onClickOutside function as a trigger to close the Popover content the onFocusOutside handler is not needed

Closes #877