fix: focus manager use with grid manager (#1228)
fix: In this change, we fix an issue with the focus manager which prevented focus from traveling into a tabbable element positioned after a focus-manged grid inside a popover. Using a grid, with controls in its cells, inside a popover (DatePicker) caused tabbable nodes expected by the focus manager to be changed by the grid manager. This is seen in Datepicker where it was not possible to navigate to the today footer button with keyboard tabbing.